moved the dataset and neural network source files to the ml model directory
