+++ train_model.py
@@ -5,8 +5,8 @@ import math
 
 from torch.utils.tensorboard import SummaryWriter
 
-from dataset import CustomDataset
-from neural_network import NeuralNetwork
+from ml_model.dataset import CustomDataset
+from ml_model.neural_network import NeuralNetwork
 
 TRAINED_MODEL_DIR = "trained_models/"
